Adjust formats for 64 bit time_t

+++ src/auditd.c
@@ -312,12 +312,13 @@ int send_audit_event(int type, const char *str)
 	// Write event into netlink area like normal events
 	if (gettimeofday(&tv, NULL) == 0) {
 		e->reply.len = snprintf((char *)e->reply.msg.data,
-			DMSG_SIZE, "audit(%lu.%03u:%u): %s", 
-			tv.tv_sec, (unsigned)(tv.tv_usec/1000), seq_num, str);
+			DMSG_SIZE, "audit(%lld.%03u:%u): %s",
+			(long long int)tv.tv_sec, (unsigned)(tv.tv_usec/1000),
+			seq_num, str);
 	} else {
 		e->reply.len = snprintf((char *)e->reply.msg.data,
-			DMSG_SIZE, "audit(%lu.%03d:%u): %s", 
-			(unsigned long)time(NULL), 0, seq_num, str);
+			DMSG_SIZE, "audit(%lld.%03d:%u): %s",
+			(long long int)time(NULL), 0, seq_num, str);
 	}
 	// Point message at the netlink buffer like normal events
 	e->reply.message = e->reply.msg.data;
